equity-model/requirements.txt
duckdb
yfinance
requests
pyyaml
pandas
plotly
rich
